chore move swagger schema classes to a dedicated package
AM-3585
