+++ gravitee-am-management-api/gravitee-am-management-api-rest/src/main/java/io/gravitee/am/management/handlers/management/api/schemas/BulkCreateOrganizationUser.java
@@ -1,0 +1,25 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.am.management.handlers.management.api.schemas;
+
+import io.gravitee.am.management.handlers.management.api.bulk.BulkRequest;
+import io.gravitee.am.service.model.NewOrganizationUser;
+
+public class BulkCreateOrganizationUser extends BulkRequest<NewOrganizationUser> {
+    public BulkCreateOrganizationUser() {
+        super(Action.CREATE);
+    }
+}
